refactor(suite-common): redo separate selectors logic
