+++ suite-common/wallet-core/src/discovery/discoveryReducer.ts
@@ -3,7 +3,6 @@ import { Discovery, DiscoveryStatus, Timestamp } from '@suite-common/wallet-type
 import { DeviceUniquePath } from '@trezor/connect';
 
 import { discoveryActions } from './discoveryActions';
-import { DeviceRootState, selectSelectedDevice } from '../device/deviceReducer';
 
 export type DiscoveryRootState = {
     wallet: {
@@ -43,48 +42,3 @@ export const prepareDiscoveryReducer = createReducerWithExtraDeps(
         });
     },
 );
-
-export const selectDiscovery = (state: DiscoveryRootState) => state.wallet.discovery;
-
-export const selectDiscoveryByDevicePath = (state: DiscoveryRootState, path?: DeviceUniquePath) =>
-    path !== undefined ? state.wallet.discovery[path] : undefined;
-
-export const selectDiscoveryForSelectedDevice = (state: DiscoveryRootState & DeviceRootState) => {
-    const selectedDevice = selectSelectedDevice(state);
-
-    return selectDiscoveryByDevicePath(state, selectedDevice?.path);
-};
-
-export function isDiscoveryInProgress(
-    discovery?: DiscoveryStatus,
-): discovery is Exclude<
-    DiscoveryStatus,
-    { status: 'complete' } | { status: 'failed' } | { status: 'cancelled' }
-> {
-    if (!discovery) {
-        return false;
-    }
-
-    return (
-        discovery.status !== 'complete' &&
-        discovery.status !== 'failed' &&
-        discovery.status !== 'cancelled'
-    );
-}
-export const selectHasRunningDiscovery = (state: DiscoveryRootState & DeviceRootState) => {
-    const discovery = selectDiscoveryForSelectedDevice(state);
-
-    return isDiscoveryInProgress(discovery);
-};
-
-// TODO remove reexport
-export const selectHasDeviceDiscovery = selectHasRunningDiscovery;
-
-/**
- * Helper selector called from components
- * return `true` if discovery process is running/completed and `authConfirm` is required
- */
-export const selectIsDiscoveryAuthConfirmationRequired = (
-    state: DiscoveryRootState & DeviceRootState,
-    path?: DeviceUniquePath,
-) => selectDiscoveryByDevicePath(state, path)?.status === 'confirm-empty-passphrase';
