+++ suite-common/wallet-core/src/index.ts
@@ -2,6 +2,7 @@ export * from './selectors';
 export * from './accounts/accountsActions';
 export * from './accounts/accountsConstants';
 export * from './accounts/accountsMiddleware';
+export * from './accounts/accountsSelectors';
 export * from './accounts/accountsReducer';
 export * from './accounts/accountsThunks';
 export * from './blockchain/blockchainActions';
@@ -15,10 +16,12 @@ export * from './explorer/explorerSelectors';
 export * from './device/deviceActions';
 export * from './device/deviceConstants';
 export * from './device/deviceReducer';
+export * from './device/deviceSelectors';
 export * from './device/deviceThunks';
 export * from './discovery/discoveryActions';
 export * from './discovery/discoveryReducer';
 export * from './discovery/discoveryThunks';
+export * from './discovery/discoverySelectors';
 export * from './discovery/passphraseUtils';
 export * from './fees/feesActions';
 export * from './fees/feesReducer';
@@ -33,6 +36,7 @@ export * from './send/composeCancelTransaction/composeCancelTransactionThunk';
 export * from './send/sendFormActions';
 export * from './send/sendFormReducer';
 export * from './send/sendFormThunks';
+export * from './send/sendFormSelectors';
 export * from './send/sendFormTypes';
 export * from './stake/stakeActions';
 export * from './stake/stakeConstants';
@@ -44,6 +48,7 @@ export * from './stake/stakeTypes';
 export * from './transactions/transactionsActions';
 export * from './transactions/transactionsReducer';
 export * from './transactions/transactionsThunks';
+export * from './transactions/transactionsSelectors';
 export * from './settings/walletSettingsActions';
 export * from './settings/walletSettingsConstants';
 export * from './settings/walletSettingsReducer';
